Add the default sub-command.
Putting the default file in edit/create was confusing, and made it difficult to find.
Having a dedicated sub-command consolidates creation and editing.

WARNING: The old behavior of edit/create default do not work as they used to. They will now
create a profile called default, but it is not the default profile.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>