Add a feature sub-command
This is functionality that has been needed for a while. Its absence was because, unlike profiles, features are used globally. This means regular users should not be able to modify them, since they could affect the sandboxes of other users.

To get past this, we simply use pkcheck to ensure the running user would have the privilege to directly modify antimony's install to begin with.

This also required abstracting the profile's edit function.

Some bug fixes as well, particularly the fact that we actually needed the default handling keyword for inheriting the default profile. This means that the prior warning is no longer true. default is still a reserved keyword for edit/create, and operates identically to the default sub-command (Backwards compatability, I guess)

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>