Integration features
Integration now adds a universal native launch option, allowing users to easily launch the application without a sandbox through the desktop entry. This is done as a Desktop Action.
Refresh now supports (re)integration on top of refreshing the SOF. This is useful for updating Antimony, where the integration spec may have changed, to automatically refresh the existing desktop files.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>