Package support.
Antimony has also had the hypothetical ability to create a standalone sandbox, one that contained all the required files for a given binary, that could then be sent to other computers. Now, the package subcommand will produce a zip file that can be moved to different computers, and run. Antimony supports loading these .sb archives, and will function even if the application in question is not installed on the computer, or if dependencies are missing.

This feature introduced a breaking change, where files.system has been divided into files.platform and files.resources. Platform system files are files exclusive to the running machine, such as ephemeral folders like /sys, as well as locales and fonts. Resource system files are files that are part of a program or library, such as /usr/share/glycin. Outside of packages, the only change is that platform files are --try-bind, whereas resource files are --bind; this means that if a resource file is missing, the sandbox will abort.

All preinstalled profiles have been updated for these changes, but user-profiles may need to be modified.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>