Graceful refreshing
The prior refresh logic merely deleted the entire cache folder. This worked, but caused running instances to sometimes misbehave. For programs, nothing happened, but others could cause unexplained behavior. To resolve this, Antimony keeps track of instances not only in the runtime folder, but also in the cache folder. This allows for Antimony to associate Instances and Applications (Which before were distinct things). With this, refresh can now check if a given profile has running instances, and create a deferred cache instead. New instances will used this deferred cache, and when all instances of the original close, the deferred with seamlessly replace the original.

This work also brought proper signal handling, thanks to a new wait_for_signal() function in the Spawn crate. Antimony will now gracefully cleanup when hit with SIGTERM or SIGINT, which is critical now that proper cleanup is required to manage instances (This was technically a bug since now, but hasn't been that important since the only place residual instance information would be kept is in /run, which recycles on reboot)

Cleaned up imports.