Hooks and Import/Export
Antimony now supports pre and post hooks on sandbox execution. This allows for host binaries to prepare/cleanup the sandbox outside of Antimony's usual procedures. Pre-Hooks can be configured to run alongside the sandbox, terminating when the sandbox does.

User Profiles can be exported and imported via the dedicated commands.

Added support for reset to remove identical user profiles when no profile is provided.

Spawn now explicitly sets variables passed to it, even if preserve_env is enabled, so that environment variables can be added/changed to the environment

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>