Sandbox Arguments
Zypak needed an extensible way to inject the wrapper before the program launches. This introduced the sandbox_args option for Profiles and Features. Essentially, they can either be cooperative, like Zypak, or exclusive, like Debug Shell and Stats (Which can now be moved out of dedicated Subcommands, and into features).

Zypak also added the need for conditional features. Not all systems will have Zypak instead (In fact, few of them will), so there neded to be a way to ensure Zypak was actually installed, especially since it hooks the command line to inject itself. The conditional field in Features presents a simple shell command that can be evaluated to see if the feature can be run. This adds flexibility, so features can perform as complicated or as simple a check as needed.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>