Way too many changes for a single commit ;)
Proper support for non-setuid/hardlink configurations. Prior efforts weren't very good. Now, Antimony works exactly as expected. It does this by abstracting AT_HOME, which can contain read-only information, to a new CACHE_DIR, which defaults to AT_HOME, but will fallbackc to /tmp/antimony if it can't write where it's supposed to. In cases of hard-link failure, Antimony will create copies in /tmp, then harldink from a shared location (Borrowed from SB), so that profiles can share copies are reduce duplication. This allows for interface level testing benchmarks (Since we can run the immediate artifacts, now), stracing antimony itself, not just using strace within the profile, and more advanced compilations (See below)

PGO and BOLT support now works. It seems BOLT is actually worse than just a release binary, but PGO has marginal benefits, although probably not worth the added complexity and compile time.

Finally renamed aux to shared, that way we can store the repo on Windows.

The user crate now has a run_as macro, which encapsulates saving the current state, executing a block under a specific mode, then reverting back. This fixes potential bugs, and cleans up the code.

Configurations can now be embedded into binary desktops with the --create-desktop flag.

Profile homes can explicitly provide a mount path.

Profiles can provide a parent hook, which operates differently than pre/post. A parent hook takes control of the sandbox. It effectively acts as a pre-attach hook, but there can only be a single one, and if the parent hook dies, the sandbox dies with it. This can be useful for, say, attaching syncthingtray-qt6 to the syncthing profile. It opens the tray icon on startup, and kills syncthing when the tray exists

Tracing now supports dedicated arguments to pass to strace via --trace-args, with passthrough operating as it does in run, and passing it to the sandbox directly.

Strace now uses -Y, so it can resolve PIDs to names.

New profiles: Syncthingtray and SSHD. The latter needs some extra configuration, but allows for running a user-level SSH daemon, sandboxed with whatever is needed in that environment.