Re-write the benchmarker in Rust.
This lets us use Clap for argument parsing, which makes it a lot easier to use.

It also now has time-travel support, but much better than the old SB version. It uses bind mounts to ensure the configuration is not only compatible with the checkout, but that user configuration and SECCOMP are temporarily cleared to ensure that each version uses the same feature set.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>