Various improvmenets
The library and binary fabricator have been improved to work for user-applications, as well as for different library configurations.

run_as has been renamed try_run_as, with run_as now being an identical macro that uses unwrap, for situations where Results cannot be used. This is in tandem with a general overhaul of how SetUID is managed within Antimony. It now defaults to Effective, and elevates to Real when needed. This makes it a lot easier to know what mode certain functions are in at any given time.

Spawn's handle::wait now takes an Optional timeout. This is used in the binary fabricator to ensure bash calls do not hang setup. However, the implementation just uses a busy-loop, so it should be used sparingly.