Work on a new binary parser
The existing parser isn't great, but it does the job. The antimony-dumper simply dumps the arguments to syscalls, which was supposed to be more reliable, and faster, than the original parser. Unfortunately, its neither. However, this is a neat application of the SECCOMP notifier, and there is a chance it might be improved later on.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>