Allow hooks in features
This allows the encryption example mentions in docs/Hooks.md to actually just be implemented as a feature.

Fixed a bug where pre-hooks ran under the effective user, which could cause a deadlock where a panic in the hook processing could cause cleanup to run under effective, but real is needed to cleanup the sandbox. Now, hooks are run within the same run_as as the main loop (This also enforces user-mode hooks, as if that weren't already enforced already)

Features can also now define the flags in IPC, though they have special rules regarding overriding.