Incorporate logging into Spawn.
Both Standard Input and Output can be redirected to an anonymous Stream with an associated log level. This Stream is not owned by the Handle, but a separate worker thread, who will print each line through the logger at the specified log level. If the current log level is higher than the provided, no pipe will be created, and the stream will be treated like StreamMode::Share
Coupled with the notify crate, our proxy and sandbox will not automatically log errors through the notify framework (Though the sandbox logs at the warn level because some applications write non-critical messages to standard error

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>