More speed and features
Add a new temp crate to remove tempfile.
Dry now works explicitly with bash, to avoid symlinks.
The benchmarker now accepts a minimum amount of runs, an option to pause for inspection, and the ability to benchmark the system refresh.
The spawner is now thread safe for more options, including the cache and seccomp
The dumper is now single-threaded besides the kernel FD to avoid deadlocking conditions.
The SECCOMP Pool Connection is now behind an Option, so we can account for failures in connection (Like permission), and act accordingly without causing a panic.
Actually use bin's ParseReturn caching. We just... didn't before. Now, the cache properly works for new parse calls, and for the entire instance via a new bin.cmd in the instance folder.
The binary fabricator now directly creates a /usr/bin to mount, akin to an SOF. This reduces command line complexity and profile juggling between fabricators.
Fabricators now used a shared FabInfo structure to reduce the amount of arguments.
add_sof can now be supplied a destination path so we can forgo symlinking and just instantiate libaries/binaries as they appear to the sandbox.
Remove flatten where possible.
The proxy is now initialized on a separate thread from the rest of the fabricators, speeding up performance considerably.
A few loops within lib and bin fabricators needed to be done single-threaded to avoid strange deadlocks.
SECCOMP setup will simply refuse to install a filter if it could not connect to the Pool, avoiding panics.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>