Reentrant User
The User synchronization primitive now supports reentrance. If the lock is held by the current thread, we can simply skip acquiring the lock, since we know that all other threads will not change the state, and due to the nature of the run_as we will restore the mode to what the thread expects prior to the call.
The sync feature for user has been consolidated. All run_as functions use the sync primitive.
try_run_as has been removed in favor of run_as simply returning a Result for its own functions, to which the body could return its own, or not. This allows for differentiating between user errors, and block errors.
as_effective and as_real have been created as shorthands for running blocks under a particular user.
fork() doesn't like you doing things in the child, which we had done to try and improve performance, but which was causing deadlocks. More investigation is needed to see if the wildcards and collection deadlocks still persist, but they seem to have been vanquished.
Because we are now running the proxy under a new thread, there is no longer any guarantee what the current operating mode may be. This requires using run_as for every operation that requires privilege for either mode; this does not incur any performance hit, but requires some time to weed out all the cases. Most seem to have been fixed.

Signed-off-by: kkernick <90726554+kkernick@users.noreply.github.com>